Refactor package directories
