+++ cli/flags_test.go
