+++ commands/info.go
@@ -1,4 +1,4 @@
-package client
+package commands
 
 import (
 	"fmt"
